fix: keep mediator compatibility contracts scoped
