+++ src/Domain/Common/MediatorCompatibility/MediatR/PreProcessorContracts.cs
@@ -6,11 +6,3 @@ namespace MediatR.Pipeline
         Task Process(TRequest request, CancellationToken cancellationToken);
     }
 }
-
-namespace Mediator.Pipeline
-{
-    public interface IRequestPreProcessor<in TRequest> : global::MediatR.Pipeline.IRequestPreProcessor<TRequest>
-        where TRequest : notnull
-    {
-    }
-}
