+++ tests/Application.UnitTests/Infrastructure/MediatorCompatibility/ScopedMediatorTests.cs
@@ -6,10 +6,11 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using CleanArchitecture.Blazor.Infrastructure.Services.MediatorWrapper;
+using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
-namespace CleanArchitecture.Blazor.Application.UnitTests.Infrastructure.MediatorCompatibility;
+namespace CleanArchitecture.Blazor.Application.UnitTests.Infrastructure.MediatorWrapper;
 
 public class ScopedMediatorTests
 {
@@ -25,7 +26,7 @@ public class ScopedMediatorTests
     {
         var services = new ServiceCollection();
         services.AddScoped<ScopeProbe>();
-        services.AddScoped<Mediator.IMediator, FakeMediator>();
+        services.AddScoped<IMediator, FakeMediator>();
 
         using ServiceProvider provider = services.BuildServiceProvider(new ServiceProviderOptions
         {
@@ -47,7 +48,7 @@ public class ScopedMediatorTests
         Assert.That(FakeMediator.ResolvedProbeIds, Is.EquivalentTo(ScopeProbe.CreatedIds));
     }
 
-    private sealed record PingQuery : Mediator.IRequest<string>;
+    private sealed record PingQuery : IRequest<string>;
 
     private sealed class ScopeProbe : IDisposable
     {
@@ -80,7 +81,7 @@ public class ScopedMediatorTests
         }
     }
 
-    private sealed class FakeMediator : Mediator.IMediator
+    private sealed class FakeMediator : IMediator
     {
         public static int SendCallCount { get; private set; }
 
@@ -99,7 +100,7 @@ public class ScopedMediatorTests
             ResolvedProbeIds.Clear();
         }
 
-        public Task<TResponse> Send<TResponse>(Mediator.IRequest<TResponse> request, CancellationToken cancellationToken = default)
+        public Task<TResponse> Send<TResponse>(IRequest<TResponse> request, CancellationToken cancellationToken = default)
         {
             SendCallCount++;
             ResolvedProbeIds.Add(_scopeProbe.Id);
@@ -113,7 +114,7 @@ public class ScopedMediatorTests
         }
 
         public Task Send<TRequest>(TRequest request, CancellationToken cancellationToken = default)
-            where TRequest : Mediator.IRequest
+            where TRequest : IRequest
         {
             throw new NotSupportedException();
         }
@@ -124,7 +125,7 @@ public class ScopedMediatorTests
         }
 
         public Task Publish<TNotification>(TNotification notification, CancellationToken cancellationToken = default)
-            where TNotification : Mediator.INotification
+            where TNotification : INotification
         {
             throw new NotSupportedException();
         }
@@ -135,7 +136,7 @@ public class ScopedMediatorTests
         }
 
         public async IAsyncEnumerable<TResponse> CreateStream<TResponse>(
-            Mediator.IStreamRequest<TResponse> request,
+            IStreamRequest<TResponse> request,
             [EnumeratorCancellation] CancellationToken cancellationToken = default)
         {
             await Task.CompletedTask;
